+++ itaca/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities: unit conversion, validation, memory helpers."""